Move swipeEvents and zoom up a hierarchy. Listen for wheel events on the webview and check the settings for if swipe is disabled.
