+++ pages/settings/index.html
@@ -46,12 +46,12 @@
 	</div>
 
     <div class="spacer"></div>
-    <div class="container padding" id="appearance-settings-container">
+    <div class="container padding" id="swipe-settings-container">
         <h3>Swipe Navigation</h3>
 
         <div class="setting-section">
             <input type="checkbox" id="checkbox-swipe-navigation" />
-            <label for="checkbox-swipe-navigation">Disable swipe for page navigation.</label>
+            <label for="checkbox-swipe-navigation">Disable swipe for navigating back and forth between pages.</label>
         </div>
     </div>
 
@@ -82,7 +82,6 @@
 	<script src="../../js/util/defaultKeyMap.js"></script>
 	<script src="../../js/util/settings.js"></script>
 	<script src="../../js/util/searchEngine.js"></script>
-    <script src="../../js/util/swipeEvents.js"></script>
 	<script src="../../js/util/theme.js"></script>
 	<script src="settings.js"></script>
 </body>
